// panel/src/website/queries/types/Website.ts
/* tslint:disable */
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL query operation: Website
// ====================================================

export interface Website_website {
  __typename: "Website";
  name: string;
  description: string;
  domain: string;
}

export interface Website {
  website: Website_website;
}